feat: support preview releases in publish script
The publish script has been updated to support a new 'preview' version.
When releasing a preview version, the following changes occur:

- A branch name must be provided, and the release will be built from that branch.
- The npm package is published with a '--tag=preview' flag.
- The following steps are skipped:
  - Pushing changes to the master branch.
  - Creating a GitHub release.
  - Generating firepit artifacts.
  - Generating a Docker image.

+++ scripts/publish.sh
@@ -2,21 +2,28 @@
 set -e
 
 printusage() {
-  echo "publish.sh <version>"
+  echo "publish.sh <version> [branch]"
   echo "REPOSITORY_ORG and REPOSITORY_NAME should be set in the environment."
   echo "e.g. REPOSITORY_ORG=user, REPOSITORY_NAME=repo"
   echo ""
   echo "Arguments:"
-  echo "  version: 'patch', 'minor', 'major', or 'artifactsOnly'"
+  echo "  version: 'patch', 'minor', 'major', 'artifactsOnly', or 'preview'"
+  echo "  branch: required if version is 'preview'"
 }
 
 VERSION=$1
+BRANCH=$2
 if [[ $VERSION == "" ]]; then
   printusage
   exit 1
 elif [[ $VERSION == "artifactsOnly" ]]; then
   echo "Skipping npm package publish since VERSION is artifactsOnly."
   exit 0
+elif [[ $VERSION == "preview" ]]; then
+  if [[ $BRANCH == "" ]]; then
+    printusage
+    exit 1
+  fi
 elif [[ ! ($VERSION == "patch" || $VERSION == "minor" || $VERSION == "major") ]]; then
   printusage
   exit 1
@@ -61,6 +68,11 @@ echo "Moved to temporary directory."
 echo "Cloning repository..."
 git clone "git@github.com:${REPOSITORY_ORG}/${REPOSITORY_NAME}.git"
 cd "${REPOSITORY_NAME}"
+if [[ $VERSION == "preview" ]]; then
+  echo "Checking out branch $BRANCH..."
+  git checkout "$BRANCH"
+  echo "Checked out branch $BRANCH."
+fi
 echo "Cloned repository."
 
 echo "Making sure there is a changelog..."
@@ -78,10 +90,17 @@ echo "Running tests..."
 npm test
 echo "Ran tests."
 
-echo "Making a $VERSION version..."
-npm version $VERSION
-NEW_VERSION=$(jq -r ".version" package.json)
-echo "Made a $VERSION version."
+if [[ $VERSION == "preview" ]]; then
+  echo "Making a preview version..."
+  npm version prerelease --preid=preview
+  NEW_VERSION=$(jq -r ".version" package.json)
+  echo "Made a preview version."
+else
+  echo "Making a $VERSION version..."
+  npm version $VERSION
+  NEW_VERSION=$(jq -r ".version" package.json)
+  echo "Made a $VERSION version."
+fi
 
 echo "Making the release notes..."
 RELEASE_NOTES_FILE=$(mktemp)
@@ -92,23 +111,30 @@ cat CHANGELOG.md >> "${RELEASE_NOTES_FILE}"
 echo "Made the release notes."
 
 echo "Publishing to npm..."
-npx clean-publish@5.0.0 --before-script ./scripts/clean-shrinkwrap.sh
+if [[ $VERSION == "preview" ]]; then
+  # Note: we publish with --tag=preview so that this does not become the "latest" version
+  npx clean-publish@5.0.0 --before-script ./scripts/clean-shrinkwrap.sh --tag=preview
+else
+  npx clean-publish@5.0.0 --before-script ./scripts/clean-shrinkwrap.sh
+fi
 echo "Published to npm."
 
-echo "Updating package-lock.json for Docker image..."
-npm --prefix ./scripts/publish/firebase-docker-image install
-echo "Updated package-lock.json for Docker image."
+if [[ $VERSION != "preview" ]]; then
+  echo "Updating package-lock.json for Docker image..."
+  npm --prefix ./scripts/publish/firebase-docker-image install
+  echo "Updated package-lock.json for Docker image."
 
-echo "Cleaning up release notes..."
-rm CHANGELOG.md
-touch CHANGELOG.md
-git commit -m "[firebase-release] Removed change log and reset repo after ${NEW_VERSION} release" CHANGELOG.md scripts/publish/firebase-docker-image/package-lock.json
-echo "Cleaned up release notes."
+  echo "Cleaning up release notes..."
+  rm CHANGELOG.md
+  touch CHANGELOG.md
+  git commit -m "[firebase-release] Removed change log and reset repo after ${NEW_VERSION} release" CHANGELOG.md scripts/publish/firebase-docker-image/package-lock.json
+  echo "Cleaned up release notes."
 
-echo "Pushing to GitHub..."
-git push origin master --tags
-echo "Pushed to GitHub."
+  echo "Pushing to GitHub..."
+  git push origin master --tags
+  echo "Pushed to GitHub."
 
-echo "Publishing release notes..."
-hub release create --file "${RELEASE_NOTES_FILE}" "v${NEW_VERSION}"
-echo "Published release notes."
+  echo "Publishing release notes..."
+  hub release create --file "${RELEASE_NOTES_FILE}" "v${NEW_VERSION}"
+  echo "Published release notes."
+fi
